requirements.txt
gunicorn>=19.6,<19.7 # server
psycopg2>=2.6.2,<2.7 # postgres bindings
Django>=1.11,<2.0
wagtail>=1.10,<1.11
markdown>=2.6.7,<2.7
requests>=2.13,<2.14
django-compressor>=2.1,<2.2
django-supervisor>=0.4.0,<0.5
django-taggit-templatetags2>=1.6.1,<1.7
django-redis>=4.7,<4.8

